build: fix code to resolve compiler warnings
This patch fixes the code so that when it is compiled with clang (v15)
with the same flags used for compiling LLVM (i.e. flags set in LLVM's
HandleLLVMOptions.cmake), the compiler no longer emits warnings.
Specifically, this patch addresses the following problems:

 - Fixes or removes unused variables (`-Wunused-variables` and
 `-Wunused-but-set-parameter`)

 - Specifies template type explicitly (`-Wctad-maybe-unsupported`)

 - Uses consistent (`struct` versus `class`) type declarations
 (`-Wmismatched-types`)

 - Fixes order of field initialization in class constructor
 (`-Wreorder-ctor`)

 - Fixes comparison between signed and unsigned types (`-Wsign-compare`)

 - Removes redundant (and possibly dangerous) node initialization in
 MLIR dataflow analysis

 - Replaces C-style casts with `static_cast`, so that such casts are
 more easily searchable (to pin down possible bugs)

 - Marks overriden methods with `override` (`-Wsuggest-override`)

To ensure that we don't regress, this patch also adds LLVM's
HandleLLVMOptions to the top-level CMakeLists.txt.

+++ lib/Conversion/TritonGPUToLLVM/Utility.h
@@ -241,7 +241,7 @@ struct SharedMemoryObject {
   }
 
   Value getCSwizzleOffset(int order) const {
-    assert(order >= 0 && order < strides.size());
+    assert(order >= 0 && order < static_cast<int>(strides.size()));
     return offsets[order];
   }
